feat(condo): DOMA-13494 six iteration
