+++ packages/bridge/src/types/methods/index.ts
@@ -34,7 +34,7 @@ export type { RequestPaymentParams, RequestPaymentData } from './RequestPayment'
 export type { ResizeWindowParams, ResizeWindowData } from './ResizeWindow'
 export type { ShowModalWindowParams, ShowModalWindowData } from './modals/ShowModalWindow'
 export type { ShowNotificationParams, ShowNotificationData } from './ShowNotification'
-export type { ShowProgressBarParams, ShowProgressBarData, TrackableCondoTaskSchemaName } from './progress-bars/ShowProgressBar'
+export type { ShowProgressBarParams, ShowProgressBarData } from './progress-bars/ShowProgressBar'
 export type { UpdateModalWindowParams, UpdateModalWindowData } from './modals/UpdateModalWindow'
 export type { UpdateProgressBarParams, UpdateProgressBarData } from './progress-bars/UpdateProgressBar'
 export type { SetPageActionsParams, SetPageActionsData } from './SetPageActions'
