+++ packages/ui/src/components/Table/components/SelectionCheckbox.tsx
@@ -1,7 +1,8 @@
-import type { CheckboxChangeEvent } from 'antd/lib/checkbox'
 import React from 'react'
 
 import { Checkbox, Tooltip } from '@open-condo/ui/src'
+
+import type { CheckboxChangeEvent } from 'antd/lib/checkbox'
 
 
 type SelectionCheckboxProps = {
